+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+- Option to run workflows inside Docker containers using the `docker_image`
+  parameter on `RunJobRequest`.
+- Workflow containers now receive all variables from local settings and secrets
+  files as environment variables.
+
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).